+++ pygount/__init__.py
@@ -3,14 +3,17 @@ Pygount counts lines of source code using pygments lexers.
 """
 # Copyright (c) 2016-2020, Thomas Aglassinger.
 # All rights reserved. Distributed under the BSD License.
-from pygount.analysis import encoding_for, DuplicatePool, SourceAnalysis, SourceScanner, SourceState, source_analysis
-from pygount.common import __version__, Error, OptionError
+from .analysis import encoding_for, DuplicatePool, SourceAnalysis, SourceScanner, SourceState, source_analysis
+from .common import __version__, Error, OptionError
+from .summary import LanguageSummary, ProjectSummary
 
 __all__ = [
     "encoding_for",
     "DuplicatePool",
     "Error",
+    "LanguageSummary",
     "OptionError",
+    "ProjectSummary",
     "SourceAnalysis",
     "SourceScanner",
     "SourceState",